Initial commit
Currently building a basic markdown to html static site generator thing.
Right now it just converts markdown titles to html-style <h%d> tags.

The idea for this is to be some kind of recepie website a-la the now
(unfortunately) defunct based.cooking (not that I'm some kind of
master chef, I just kinda need something to do to stop my mind
from wandering rn, not that that's a good thing).

Enjoy future archeologists!